requirements.txt
amqp==5.0.6
arrow==0.14.7
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dj-rest-auth==2.1.11
Django==3.2.6
django-celery-beat==2.2.1
django-filter==2.4.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
flower==1.0.0
humanize==3.11.0
ics==0.7
idna==3.2
kombu==5.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
TatSu==5.6.1
tornado==6.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
